requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
filelock==3.18.0
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
griffe==1.7.3
h11==0.16.0
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ImageHash==4.3.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.1.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mcp==1.9.4
missingno==0.5.2
multimethod==2.0
narwhals==1.42.1
networkx==3.5
numpy==2.3.0
openai==1.86.0
openai-agents==0.0.17
packaging==24.2
pandas==2.3.0
pandas-profiling==3.2.0
phik==0.12.4
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.1
pyarrow==19.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
s3transfer==0.13.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
sse-starlette==2.3.6
starlette==0.47.0
streamlit==1.45.1
streamlit-ace==0.1.1
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
visions==0.7.4
watchdog==6.0.0
